refactor(webapp): group currency form files under form/ subdir
