+++ frontend/projects/webapp/src/app/core/currency/form/amount-form.types.ts
