+++ frontend/projects/webapp/src/app/core/currency/form/conversion-form.schema.ts
